+++ shared/elasticsearch/src/esClientUtils.js
@@ -1,15 +1,13 @@
-const { logger } = require("@shared/utils");
-
 const { analyzer, char_filter, filter, tokenizer } = require("./analysis");
 
 async function createIndex({ client, indexName, mappings }) {
-  const { body } = await client.indices.exists({ index: indexName });
+  const body = await client.indices.exists({ index: indexName });
   if (body) {
     try {
       await client.indices.delete({ index: indexName });
-      logger.info(`Index ${indexName} deleted.`);
+      console.info(`Index ${indexName} deleted.`);
     } catch (error) {
-      logger.error("index delete", error);
+      console.error("index delete", error);
     }
   }
   try {
@@ -31,15 +29,15 @@ async function createIndex({ client, indexName, mappings }) {
       },
       index: indexName,
     });
-    logger.info(`Index ${indexName} created.`);
+    console.info(`Index ${indexName} created.`);
   } catch (error) {
-    logger.error("index create", error);
+    console.error("index create", error);
   }
 }
 
 async function version({ client }) {
-  const { body } = await client.info();
-  logger.info(body.version.number);
+  const body = await client.info();
+  console.info(body.version.number);
 }
 
 async function bulkIndexDocuments({ client, indexName, documents }) {
@@ -63,15 +61,13 @@ async function bulkIndexDocuments({ client, indexName, documents }) {
       ),
       index: indexName,
     });
-    if (resp.body.errors) {
-      const errorDocs = resp.body.items.filter(
-        (item) => item.index.status !== 201
-      );
+    if (resp.errors) {
+      const errorDocs = resp.items.filter((item) => item.index.status !== 201);
       throw new Error(`Error during indexing ${JSON.stringify(errorDocs)}`);
     }
-    logger.info(`Index ${documents.length} documents.`);
+    console.info(`Index ${documents.length} documents.`);
   } catch (error) {
-    logger.error(`Failed to index documents (${error})`);
+    console.error(`Failed to index documents (${error})`);
     throw error;
   }
 }
@@ -82,14 +78,14 @@ async function indexDocumentsBatched({
   documents,
   size = 500,
 }) {
-  logger.info(`Loaded ${documents.length} documents`);
+  console.info(`Loaded ${documents.length} documents`);
   for (const chunk of chunks(documents, size)) {
     await bulkIndexDocuments({ client, documents: chunk, indexName });
   }
 }
 
 async function deleteOldIndex({ client, patterns, timestamp }) {
-  const { body: indices } = await client.cat.indices({ format: "json" });
+  const indices = await client.cat.indices({ format: "json" });
 
   const IndicesToDelete = getIndicesToDelete(patterns, timestamp, indices);
   const pIndicesToDelete = IndicesToDelete.map(({ index }) =>
@@ -97,7 +93,7 @@ async function deleteOldIndex({ client, patterns, timestamp }) {
   );
 
   return Promise.all(pIndicesToDelete).then(() => {
-    logger.info(
+    console.info(
       `Remove ${pIndicesToDelete.length} old indices (${JSON.stringify(
         IndicesToDelete
       )})`