+++ targets/export-elasticsearch/src/ingester/index.ts
@@ -2,5 +2,4 @@ export * from "./breadcrumbs";
 export * from "./cdtnDocuments";
 export * from "./context";
 export * from "./ingest";
-export * from "./monolog";
 export * from "./suggestion";