+++ targets/frontend/src/components/contributions/questions/Question.query.ts
@@ -43,6 +43,8 @@ query SelectQuestion($questionId: uuid) {
     contentAgreement
     contentLegal
     contentNotHandled
+    contentNotHandledWithoutLegal
+    contentAgreementWithoutLegal
     id
     label
   }